Moved everything to `sharedServer`
