+++ package.json
@@ -45,7 +45,7 @@
     "deploy:rules:firestore": "firebase deploy --only firestore:rules",
     "deploy:rules:storage": "firebase deploy --only storage",
     "deploy:indexes": "firebase deploy --only firestore:indexes",
-    "update-deps": "npx npm-check-updates -u --packageFile 'package.json' && npx npm-check-updates -u --packageFile 'packages/*/package.json' --reject '@conductor/shared,@conductor/shared-client' && yarn ci",
+    "update-deps": "npx npm-check-updates -u --packageFile 'package.json' && npx npm-check-updates -u --packageFile 'packages/*/package.json' --reject '@conductor/shared,@conductor/shared-client,@conductor/shared-server' && yarn ci",
     "test": "yarn workspaces foreach --all -p -t run test",
     "test:shared": "yarn workspace @conductor/shared test",
     "test:shared:watch": "yarn workspace @conductor/shared test:watch",
