+++ packages/functions/package.json
@@ -20,12 +20,12 @@
     "validate": "yarn run validate:lint && yarn run validate:types && yarn run validate:format",
     "validate:lint": "eslint src --report-unused-disable-directives --max-warnings 0",
     "validate:lint:fix": "eslint src --fix",
-    "validate:types": "yarn workspace @conductor/shared build && tsc --noEmit",
+    "validate:types": "yarn workspace @conductor/shared-server build && tsc --noEmit",
     "validate:format": "prettier --check 'src/**/*.{js,ts,json}'",
     "format": "prettier --write 'src/**/*.{js,ts,json}'",
     "clean": "rm -rf dist/",
     "ci": "rm -rf node_modules/ && yarn install",
-    "build": "yarn run clean && yarn workspace @conductor/shared build && esbuild src/index.ts --bundle --platform=node --target=node22 --outfile=dist/index.js --external:firebase-admin --external:firebase-functions",
+    "build": "yarn run clean && yarn workspace @conductor/shared-server build && esbuild src/index.ts --bundle --platform=node --target=node22 --outfile=dist/index.js --external:firebase-admin --external:firebase-functions",
     "serve": "yarn run build && firebase emulators:start --only functions",
     "shell": "yarn run build && firebase functions:shell",
     "start": "yarn run shell",
